feat(#84): [Tilak|Pooja] use new uri format for qr code in ios
