+++ ios/Openid4vpBle/WalletModule.swift
@@ -14,7 +14,7 @@ class WalletModule: RCTEventEmitter {
 
 
     @objc func noop() -> Void {}
-
+    
     @objc func startConnection(_ uri: String) {
         let openId4VpURI = OpenId4vpURI(uri: uri)
         print("startConnection->uri::\(uri)")
@@ -23,12 +23,7 @@ class WalletModule: RCTEventEmitter {
             return
         }
         
-        let advPayload = openId4VpURI.extractPayload()
-        if advPayload == "" {
-            ErrorHandler.sharedInstance.handleException(type: .walletException, error: .invalidURIException)
-            return
-        }
-        
+        let advPayload = getAdvPayload(openId4VpURI)
         wallet = Wallet()
         wallet?.setAdvIdentifier(identifier: advPayload)
         wallet?.startScanning()
@@ -79,5 +74,10 @@ class WalletModule: RCTEventEmitter {
         wallet?.handleDestroyConnection(isSelfDisconnect: false)
         EventEmitter.sharedInstance.emitErrorEvent(message: message, code: code)
     }
-
+    
+    
+    fileprivate func getAdvPayload(_ openId4VpURI: OpenId4vpURI) -> Data {
+        return (openId4VpURI.getName()! + "_").data(using: .utf8)!
+        + hexStringToData(string: String(openId4VpURI.getHexPK()!.prefix(10)))
+    }
 }
